Refactor ECEF path interpolation
